% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aesCoeff}
\alias{aesCoeff}
\title{Coefficients for Partition Function Models.}
\format{
A data frame with 46 observations of 18 variables:

\describe{
\item{element}{name of element}
\item{symbol}{element's atomic symbol}
\item{at_no}{element's atomic number}
\item{deGalan_a}{first coefficient of deGalan model}
\item{deGalan_b}{second coefficient of deGalan model}
\item{deGalan_c}{third coefficient of deGalan model}
\item{deGalan_d}{fourth coefficient of deGalan model}
\item{deGalan_e}{fifth coefficient of deGalan model}
\item{deGalan_f}{sixth coefficient of deGalan model}
\item{tamaki_a}{first coefficient of Tamaki model}
\item{tamaki_b}{second coefficient of Tamaki model}
\item{tamaki_c}{third coefficient of Tamaki model}
\item{irwin_a0}{first coefficient of Irwin model}
\item{irwin_a1}{second coefficient of Irwin model}
\item{irwin_a2}{third coefficient of Irwin model}
\item{irwin_a3}{fourth coefficient of Irwin model}
\item{irwin_a4}{fifth coefficient of Irwin model}
\item{irwin_a5}{sixth coefficient of Irwin model}
}
}
\source{
Coefficients are adapted from the following sources: deGalan, L.; Smith, R.; Winefordner, J. D. "The Electronic Partition Functions of Atoms and Ions Between 1500 K and 7000 K" \emph{Spectrochemica Acta}, \strong{1968}, \emph{23B}, 521-525; Tamaki, S.; Kuroda, T. "The Electronic Partition Functions of Atoms and Ions Between 7000 K and 12000 K" \emph{Spectrochimca Acta} \strong{1987}, \emph{42B}, 1105-1111; Irwin, A. W. "Polynomial Partition Function Approximations of 344 Atomic and Molecular Species" \emph{The Astrophysical Journal Supplement Series}, \strong{1981}, \emph{45}, 621-633.
}
\usage{
aesCoeff
}
\description{
Coefficients for the Irwin, deGalan, and Tamaki models for predicting an element's temperature-dependent partition function.
}
\keyword{datasets}
