% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomicEmission-package.R
\docType{package}
\name{atomicEmission-package}
\alias{atomicEmission-package}
\title{Atomic Emission Spectroscopy}
\description{
A collection of functions for simulating and visualizing
atomic emission spectra between 390 nm and 700 nm. Atomic
emission spectra are simulated using one of three models for
estimating an element's temperature-dependent partition
function. Emission spectra are displayed as either peaks or as
lines, and the corresponding atomic absorption line spectrum
may be viewed.
}
\author{
David T. Harvey

Maintainer: David T. Harvey \email{harvey@depauw.edu}
}
