% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyPeaks.R
\name{identifyPeaks}
\alias{identifyPeaks}
\title{Identify Peaks in AES Spectrum}
\usage{
identifyPeaks(file, signal_type = c("relative", "absolute"),
  percent_max = 1, num_peaks = 5, sortby = c("intensity", "wavelength"))
}
\arguments{
\item{file}{The name of the object, created using
\code{simulateAES}, that contains the results of a
simulation.}

\item{signal_type}{How to report the emission intensity; one
of \code{"relative"}, which scales the intensity to a
maximum value of 1, or \code{"absolute"}, which uses the
actual emission intensities. Defaults to \code{"relative"}.}

\item{percent_max}{The minimum intensity for which a peak is
identified, given as a percentage of the element's maximum
absolute or relative emission intensity. Defaults to 1\% of
the element's maximum absolute or relative emission
intensity.}

\item{num_peaks}{The maximum number of peaks to return.
Defaults to a table with a maximum of 5 peaks.}

\item{sortby}{How to sort the peaks; one of
\code{"intensity"}, which sorts the list of peaks from
highest to lowest intensity, or \code{"wavelength"}, which
sorts the list of peaks from shortest to longest wavelength.}
}
\value{
Returns a data frame with two columns: the wavelengths
  (in nm) of the peaks and the absolute or relative
  intensities of the peaks.
}
\description{
Creates a table of the most intense peaks in an element's
simulated atomic emission spectrum. The table is constructed
using data stored in an object created using
\code{simulateAES} and is ordered from more intense peaks to
less intense peaks.
}
\examples{
hydrogen = simulateAES(atom = "H")
hydrogen_peaks = identifyPeaks(file = hydrogen)
hydrogen_peaks
}
