% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAES.R
\name{plotAES}
\alias{plotAES}
\title{Plot Atomic Emission Spectrum}
\usage{
plotAES(file, lambda.min = 390, lambda.max = 700,
  signal_type = c("relative", "absolute"), main_title = NULL)
}
\arguments{
\item{file}{The name of an object, created by
\code{simulateAES}, that contains the results of a
simulation.}

\item{lambda.min}{The minimum wavelength (in nm) for the
x-axis; defaults to 390 nm, which is the lower limit.}

\item{lambda.max}{The maximum wavelength (in nm) for tye
x-axis; defaults to 700 nm, which is the upper limit.}

\item{signal_type}{How the emission is scaled on the y-axis;
one of \code{"relative"}, which scales the y-axis to a
maximum value of 1, or \code{"absolute"}, which scales the
y-axis using the actual emission intensities. Defaults to
\code{"relative"}.}

\item{main_title}{An optional main title. If \code{NULL}, then
a default title gives the element's atomic symbol, the
temperature, and the partition model.}
}
\value{
Returns a plot of the element's simulated atomic
  emission spectrum.
}
\description{
Plots a simulated atomic emission spectrum with the absolute or
relative emission intensity on the y-axis and wavelength on
the x-axis. The plot is constructed using data stored in an object
created using \code{simulateAES}.
}
\examples{

oxygen = simulateAES(atom = "oxygen", temperature = 4000, width = 0.2)
plotAES(file = oxygen, signal_type = "relative",
    main_title = "Element: Oxygen, Temperature: 4000")
}
