% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLines.R
\name{plotLines}
\alias{plotLines}
\title{Plot Atomic Emission Lines}
\usage{
plotLines(
  file,
  lambda.min = 390,
  lambda.max = 700,
  spectrum_type = c("emission", "absorbance"),
  emission_bg = c("white", "black"),
  scale_emission = TRUE,
  main_title = NULL
)
}
\arguments{
\item{file}{The name of an object, created by
\code{simulateAES}, that contains the results of a
simulation.}

\item{lambda.min}{The minimum wavelength (in nm) for scaling
the x-axis; defaults to 390 nm, which is the lower limit.}

\item{lambda.max}{The maximum wavelength (in nm) for scaling
the x-axis; defaults to 700 nm, which is the upper limit.}

\item{spectrum_type}{Type of line spectrum, either
\code{emission} or \code{absorbance}. Defaults to
\code{emission}.}

\item{emission_bg}{For an emission spectrum, choice of
background color as \code{white} or \code{black}. Defaults
to \code{white}.}

\item{scale_emission}{Logical value that determines whether
the transparency of an emission line is scaled to the
emission line's relative intensity. Defaults to
\code{FALSE}, which shows all emission lines regardless of
intensity; emission lines of low intensity are too faint to
see when \code{scale_emission} is set to \code{TRUE}. The
value for \code{scale_emission} is ignored when plotting the
element's absorbance spectrum.}

\item{main_title}{An optional main title. If \code{NULL}, then
a default title gives the element's atomic symbol, the
temperature, and the partition model.}
}
\value{
Returns a plot of the element's atomic emission line
  spectrum or the element's atomic absorption line spectrum.
}
\description{
Plots an element's simulated atomic emission or atomic
absorption line spectrum. The lines are taken from an object
created using \code{simulateAES}. Atomic absorption spectra
are displayed as a set of black lines superimposed on a
background that shows the spectrum of visible electromagnetic
radiation. Atomic emission spectra are displayed as set of
lines, each of which has a color that matches its wavelength,
superimosed on either a white or a black background.
}
\examples{

hydrogen = simulateAES(atom = "H")
plotLines(file = hydrogen, spectrum_type = "emission",
   emission_bg = "white", scale_emission = TRUE)
plotLines(file = hydrogen, spectrum_type = "emission",
   emission_bg = "white", scale_emission = FALSE)
plotLines(file = hydrogen, spectrum_type = "emission",
   emission_bg = "black", scale_emission = FALSE)
plotLines(file = hydrogen, spectrum_type = "absorbance",
   emission_bg = "white", scale_emission = TRUE)
}
