% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visSpec}
\alias{visSpec}
\title{Visible Spectrum.}
\format{
A data frame with 50001 observations of two variables:
  \describe{
  \item{wavelength}{wavelength, in nm, for visible
  light}
  \item{color}{rgb value for each wavelength}
  }
}
\source{
Created locally using an internal script (createVis.R)
  and called by the function plotLines.
}
\usage{
visSpec
}
\description{
Gives the rgb color for visible light with wavelengths between
300 nm and 800 nm in increments of 0.01 nm. Used to provide
the background when displaying atomic absorption line spectra.
}
\keyword{datasets}
